+++ boofuzz/web/static/js/index.js
@@ -11,16 +11,25 @@ const StringUtilities = {
 };
 
 function update_current_run_info(response) {
-    document.getElementById('current_index').textContent = response.session_info.current_index;
-    document.getElementById('num_mutations').textContent = response.session_info.num_mutations;
-    document.getElementById('current_index_element').textContent = response.session_info.current_index_element;
-    document.getElementById('num_mutations_element').textContent = response.session_info.num_mutations_element;
-    document.getElementById('current_element').textContent = response.session_info.current_element;
+    document.getElementById('current_index').textContent = response.session_info.current_index.toLocaleString();
+    document.getElementById('num_mutations').textContent = (response.session_info.num_mutations || "many").toLocaleString();
+    document.getElementById('current_index_element').textContent = response.session_info.current_index_element.toLocaleString();
+    document.getElementById('num_mutations_element').textContent = response.session_info.num_mutations_element.toLocaleString();
+    document.getElementById('current_element').textContent = response.session_info.current_element + ":";
+    document.getElementById('current_test_case_name').textContent = response.session_info.current_test_case_name;
+    document.getElementById('exec_speed').textContent = response.session_info.exec_speed.toFixed(1) + "/sec";
+    document.getElementById('run_time').textContent = response.session_info.runtime.toFixed(0) + " sec";
 
 
-    let fraction_complete_total = response.session_info.current_index / response.session_info.num_mutations;
-    document.getElementById('progress_percentage_total').textContent = progress_percentage(fraction_complete_total);
-    document.getElementById('progress_bar_total').textContent = progress_bars(fraction_complete_total);
+    if (response.session_info.num_mutations != null) {
+        let fraction_complete_total = response.session_info.current_index / response.session_info.num_mutations;
+        document.getElementById('progress_percentage_total').textContent = progress_percentage(fraction_complete_total);
+        document.getElementById('progress_bar_total').textContent = progress_bars(fraction_complete_total);
+    }
+    else {
+        document.getElementById('progress_percentage_total').textContent = "";
+        document.getElementById('progress_bar_total').textContent = "";
+    }
 
     let fraction_complete_element = response.session_info.current_index_element / response.session_info.num_mutations_element;
     document.getElementById('progress_percentage_element').textContent = progress_percentage(fraction_complete_element);
@@ -164,8 +173,8 @@ function continually_update_current_test_case_log()
 
 function progress_bars(fraction){
     return '[' +
-        StringUtilities.repeat('=', Math.round(fraction * 50)) +
-        StringUtilities.repeat(nbsp, 50 - Math.round(fraction * 50)) + ']';
+        StringUtilities.repeat('=', Math.round(Math.min(fraction, 1) * 50)) +
+        StringUtilities.repeat(nbsp, 50 - Math.round(Math.min(fraction, 1) * 50)) + ']';
 }
 
 function progress_percentage(fraction){