+++ unit_tests/test_s_repeat.py
@@ -30,4 +30,4 @@ def scenario_output_is(context, value):
 def scenario_can_render_all_mutations(context):
     mutations = list(context.req.get_mutations())
     for mutation in mutations:
-        context.req.render(MutationContext(mutation=mutation))
+        context.req.render(MutationContext(mutations=mutation))